Refactor. Completed db schema. Added soap web service for parking sensor mock.

+++ contracts/src/main/java/pl/edu/agh/soa/contracts/TicketDTO.java
@@ -12,7 +12,7 @@ import lombok.*;
 @AllArgsConstructor
 public class TicketDTO {
     @NonNull
-    private int parkingMeterId;
+    private long parkingMeterId;
     @NonNull
     private String carId;
     @NonNull
